refactor(consumable): remove the unfinished FluidConsumable placeholder
FluidConsumable was a dead-end placeholder for the IConsumable/
ConsumableTimer refactor: never instantiated outside its own
deserializeNBT, consume() always returned false, and nothing wrote the
corresponding "type":"fluid" NBT tag. The real fluid-cooling mechanism
already runs through FluidConsumptionRateCalculator/
ReactorInputFluidManagerI.extractFluids(...), so remove the dead file
rather than let it linger half-wired, and record why plus the plan to
revisit it later.

- Delete content/multiblock/controller/consumable/FluidConsumable.java
- IConsumable: comment out (not remove) the "fluid" case in
  deserializeNBT pending a possible future rework, since the mechanism
  is still unresolved rather than abandoned
- ItemConsumable: drop the redundant (int) cast in computeTimer — the
  division is already integer division, the cast was a no-op
- AUDIT_NEOFORGE.md: remove the now-resolved FluidConsumable entries
  from §1.1/§5/§7, add a "Points corrigés" log entry for the 30 August
  removal, and document a step-by-step plan in §6 for eventually
  unifying fluid consumption under IConsumable/ConsumableTimer if the
  work is picked back up

+++ src/main/java/net/nuclearteam/createnuclear/content/multiblock/controller/consumable/ItemConsumable.java
@@ -22,7 +22,7 @@ public class ItemConsumable implements IConsumable {
     @Override
     public int computeTimer(int count) {
         int base = Math.max(1, rodType.rodTimer().get());
-        return Math.max(1, (int) (base / Math.max(1, count)));
+        return Math.max(1, base / Math.max(1, count));
     }
 
     @Override
